Reproducible builds. (#23)
When run multiple times in the same month, on the same underlying
code, the output ISO will be bit-for-bit the same each time. The
build version can be faked as well, using the -b option to the
shell wrapper. The -b option is also useful for increasing the
bump version eg. 2505.2 would be the second build in May 2025.

Resolves: #20